+++ eras/conway/impl/src/Cardano/Ledger/Conway/TxBody.hs
@@ -78,7 +78,7 @@ import Cardano.Ledger.Coin (Coin (..))
 import Cardano.Ledger.Conway.Core (
   ConwayEraTxBody (..),
  )
-import Cardano.Ledger.Conway.Delegation.Certificates (ConwayDCert, transDCert)
+import Cardano.Ledger.Conway.Delegation.Certificates (ConwayDCert, toShelleyDCert)
 import Cardano.Ledger.Conway.Era (ConwayEra)
 import Cardano.Ledger.Conway.Governance (GovernanceActionInfo, Vote)
 import Cardano.Ledger.Conway.PParams ()
@@ -309,7 +309,7 @@ instance Crypto c => ShelleyEraTxBody (ConwayEra c) where
   certsTxBodyL = notSupportedInThisEraL
   {-# INLINE certsTxBodyL #-}
 
-  certsTxBodyG = getterMemoRawType (fmap transDCert . ctbrCerts)
+  certsTxBodyG = getterMemoRawType (fmap toShelleyDCert . ctbrCerts)
 
 instance Crypto c => AllegraEraTxBody (ConwayEra c) where
   {-# SPECIALIZE instance AllegraEraTxBody (ConwayEra StandardCrypto) #-}